teste-de-validacao/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
